refactor(ItemFormSection): extract json processing to own file
